Начал работы по добавлению идентификатора таба в структуру для аддонов
