+++ Addon/Page/config.php
@@ -15,6 +15,11 @@ return array(
             'sql' => 'char(15)',
             'type' => 'Ideal_Hidden'
         ),
+        'tab_ID' => array(
+            'label' => 'ID таба, в котором расположен аддон',
+            'sql' => 'int not null default 0',
+            'type' => 'Ideal_Hidden'
+        ),
         'content' => array(
             'label' => 'Текст',
             'sql' => 'mediumtext',
